+++ lib/redis-time-series.rb
@@ -9,5 +9,5 @@ require 'redis/time_series/sample'
 require 'redis/time_series'
 
 class RedisTimeSeries
-  VERSION = '0.3.0'
+  VERSION = '0.4.0'
 end